tests/wpt/tests/html/canvas/element/filters/2d.filter.layers.convolveMatrix.exceptions.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.filter.layers.convolveMatrix.exceptions</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.filter.layers.convolveMatrix.exceptions</h1>
<p class="desc">Test exceptions on CanvasFilter() convolveMatrix</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Test exceptions on CanvasFilter() convolveMatrix");
_addTest(function(canvas, ctx) {

  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'convolveMatrix'}}); });
  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'convolveMatrix', divisor: 2}}); });
  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'convolveMatrix', kernelMatrix: null}}); });
  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'convolveMatrix', kernelMatrix: 1}}); });
  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'convolveMatrix', kernelMatrix: [[1, 0], [0]]}}); });
  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'convolveMatrix', kernelMatrix: [[1, 'a'], [0]]}}); });
  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'convolveMatrix', kernelMatrix: [[1, 0], 0]}}); });
  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'convolveMatrix', kernelMatrix: [[1, 0], [0, Infinity]]}}); });
  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'convolveMatrix', kernelMatrix: []}}); });
  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'convolveMatrix', kernelMatrix: [1]}}); });
  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'convolveMatrix', kernelMatrix: [1, 2, 3, 4]}}); });
  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'convolveMatrix', kernelMatrix: [[], []]}}); });
  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'convolveMatrix', kernelMatrix: [[1, 2], []]}}); });
  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'convolveMatrix', kernelMatrix: [[], [1, 2]]}}); });
  // This should not throw an error
  ctx.beginLayer({filter:
    {name: 'convolveMatrix', kernelMatrix: [[]]}});
  ctx.endLayer();
  ctx.beginLayer({filter:
    {name: 'convolveMatrix', kernelMatrix: [[1]]}});
  ctx.endLayer();

});
</script>
